+++ server/server.js
@@ -7,6 +7,12 @@ app.get('/api/hello', (req, res) => {
   res.send({ express: 'Hello From Express' });
 });
 
+app.post('/api/log', (req, res) => {
+  console.log(req.body);
+  // TODO: get this working
+  res.send({ express: 'Hello From Express' });
+});
+
 if (process.env.NODE_ENV === 'production') {
   app.use(express.static(path.join(__dirname, 'client/build')));